refactor: penalty 도메인 분류
