--- backend/src/main/java/com/ice/studyroom/domain/membership/domain/type/PenaltyStatus.java
@@ -1,6 +1,0 @@
-package com.ice.studyroom.domain.membership.domain.type;
-
-public enum PenaltyStatus {
-	VALID,
-	EXPIRED
-}
