+++ backend/src/main/java/com/ice/studyroom/domain/penalty/application/PenaltyService.java
@@ -1,0 +1,19 @@
+package com.ice.studyroom.domain.penalty.application;
+
+import org.springframework.stereotype.Service;
+
+import com.ice.studyroom.domain.membership.domain.entity.Member;
+import com.ice.studyroom.domain.penalty.domain.type.PenaltyReasonType;
+import com.ice.studyroom.domain.penalty.infrastructure.persistence.PenaltyRepository;
+
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class PenaltyService {
+
+
+}
