+++ backend/src/main/java/com/ice/studyroom/domain/penalty/domain/type/PenaltyStatus.java
@@ -1,0 +1,6 @@
+package com.ice.studyroom.domain.penalty.domain.type;
+
+public enum PenaltyStatus {
+	VALID,
+	EXPIRED
+}
