+++ backend/src/test/java/com/ice/studyroom/domain/membership/scheduler/PenaltyUpdateSchedulerTest.java
@@ -12,11 +12,12 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
 
 import com.ice.studyroom.domain.membership.domain.entity.Member;
-import com.ice.studyroom.domain.membership.domain.entity.Penalty;
-import com.ice.studyroom.domain.membership.domain.type.PenaltyReasonType;
+import com.ice.studyroom.domain.penalty.domain.entity.Penalty;
+import com.ice.studyroom.domain.penalty.domain.type.PenaltyReasonType;
 import com.ice.studyroom.domain.membership.domain.vo.Email;
 import com.ice.studyroom.domain.membership.infrastructure.persistence.MemberRepository;
-import com.ice.studyroom.domain.membership.infrastructure.persistence.PenaltyRepository;
+import com.ice.studyroom.domain.penalty.infrastructure.persistence.PenaltyRepository;
+import com.ice.studyroom.domain.penalty.scheduler.PenaltyUpdateScheduler;
 
 import jakarta.transaction.Transactional;
